backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
gunicorn==21.2.0
razorpay==1.4.2
setuptools>=65.5.0
Flask-APScheduler==1.13.0
cloudinary==1.44.1
requests>=2.31.0